Refactor: Home, PlaceSearch 코드 정리
ViewController의 규모를 줄이기 위한 작업으로 코드를 여러 파일로 분리했습니다.
