+++ GIL/GIL/Screen/Home/HomeViewController.swift
@@ -22,11 +22,11 @@ final class HomeViewController: BaseViewController, NavigationBarHideable {
         case recentSearchPlace([PlaceData])
     }
     
-    private var interactor: HomeBusinessLogic
-    private var homeView = HomeView()
-    private var dataSource: UICollectionViewDiffableDataSource<Section, Item>!
+    var interactor: HomeBusinessLogic
+    var homeView = HomeView()
+    var dataSource: UICollectionViewDiffableDataSource<Section, Item>!
     
-    private var recentSearchCount = 0 // 최근 검색 개수를 추적하는 변수
+    var recentSearchCount = 0 // 최근 검색 개수를 추적하는 변수
     
     // MARK: - Initialization
     init(interactor: HomeBusinessLogic) {
@@ -62,29 +62,6 @@ final class HomeViewController: BaseViewController, NavigationBarHideable {
     }
 }
 
-// MARK: - UI Updates
-extension HomeViewController {
-    private func updateSnapshot(with data: [PlaceData]) {
-        recentSearchCount = data.count
-        
-        var snapshot = dataSource.snapshot()
-        snapshot.deleteItems(snapshot.itemIdentifiers(inSection: .main).filter {
-            if case .recentSearchPlace = $0 { return true }
-            return false
-        })
-        let newItems = [Item.recentSearchPlace(data)]
-        snapshot.appendItems(newItems, toSection: .main)
-        dataSource.apply(snapshot, animatingDifferences: true)
-        
-        resetLayout()
-    }
-    
-    private func resetLayout() {
-        let newLayout = createLayout()
-        homeView.mainCollectionView.setCollectionViewLayout(newLayout, animated: false)
-    }
-}
-
 // MARK: - HomeDisplayLogic
 extension HomeViewController: HomeDisplayLogic {
     func displayFetchedData(_ data: [PlaceData]) {
@@ -94,79 +71,5 @@ extension HomeViewController: HomeDisplayLogic {
     func displaySearchScreen() {
         let searchDetailVC = PlaceSearchViewController(viewModel: PlaceSearchViewModel())
         navigationController?.pushViewController(searchDetailVC, animated: true)
-    }
-}
-
-// MARK: - Collection View Setup and Layout
-extension HomeViewController {
-    private func setupCollectionView() {
-        homeView.mainCollectionView.collectionViewLayout = createLayout()
-    }
-
-    private func createLayout() -> UICollectionViewLayout {
-        return UICollectionViewCompositionalLayout { sectionIndex, layoutEnvironment in
-            let section = Section.allCases[sectionIndex]
-            switch section {
-            case .main: return self.createMainSection()
-            }
-        }
-    }
-    
-    private func createMainSection() -> NSCollectionLayoutSection {
-        let searchItemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .estimated(80))
-        let searchItem = NSCollectionLayoutItem(layoutSize: searchItemSize)
-        
-        let recentSearchItemHeight = CGFloat(50 * recentSearchCount)
-        let recentSearchItemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .estimated(recentSearchItemHeight))
-        let recentSearchItem = NSCollectionLayoutItem(layoutSize: recentSearchItemSize)
-        
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .estimated(230))
-        let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitems: [searchItem, recentSearchItem])
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.interGroupSpacing = 10
-        return section
-    }
-    
-}
-
-// MARK: - Collection View Data Source Configuration
-extension HomeViewController {
-    private func configureDataSource() {
-        dataSource = UICollectionViewDiffableDataSource<Section, Item>(collectionView: homeView.mainCollectionView) { (collectionView, indexPath, item) -> UICollectionViewCell? in
-            switch item {
-            case .search: return self.configureSearchCell(for: collectionView, at: indexPath)
-            case let .recentSearchPlace(placeData): return self.configureRecentSearchPlaceCell(for: collectionView, at: indexPath, with: placeData)
-            }
-        }
-
-        var snapshot = NSDiffableDataSourceSnapshot<Section, Item>()
-        snapshot.appendSections([.main])
-        snapshot.appendItems([.search], toSection: .main)
-        dataSource.apply(snapshot, animatingDifferences: false)
-    }
-    
-    private func configureSearchCell(
-        for collectionView: UICollectionView,
-        at indexPath: IndexPath
-    ) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: HomeSearchCollectionViewCell.reuseIdentifier, for: indexPath) as? HomeSearchCollectionViewCell else { return UICollectionViewCell() }
-        
-        cell.onSearchBarTapped = { [weak self] in
-            guard let self else { return }
-            self.interactor.performSearch()
-        }
-        return cell
-    }
-
-    private func configureRecentSearchPlaceCell(
-        for collectionView: UICollectionView,
-        at indexPath: IndexPath,
-        with data: [PlaceData]
-    ) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: HomeRecentSearchPlaceCollectionViewCell.reuseIdentifier, for: indexPath) as? HomeRecentSearchPlaceCollectionViewCell else { return UICollectionViewCell() }
-        
-        cell.configure(with: data)
-        return cell
     }
 }
