+++ GIL/GIL/Screen/PlaceSearch/Extension/PlaceSearchViewController+CollectionViewSetup.swift
@@ -1,0 +1,28 @@
+//
+//  PlaceSearchViewController+CollectionViewSetup.swift
+//  GIL
+//
+//  Created by 송우진 on 5/1/24.
+//
+
+import UIKit
+
+extension PlaceSearchViewController {
+    func setupCollectionView() {
+        placeSearchView.searchResultsCollectionView.collectionViewLayout = createLayout()
+    }
+    
+    private func createLayout() -> UICollectionViewLayout {
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .estimated(70))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1.0), heightDimension: .estimated(70))
+        let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitems: [item])
+
+        let section = NSCollectionLayoutSection(group: group)
+        section.interGroupSpacing = 5
+        section.contentInsets = NSDirectionalEdgeInsets(top: 10, leading: 16, bottom: 10, trailing: 16)
+
+        return UICollectionViewCompositionalLayout(section: section)
+    }
+}
